Initial commit for MVP version with clean and well-written code, and thoroughly tested with no bugs found
Welcome to "Catch The Egg", a fun game created by me, Nursultan Akhmetov, as part of my application for the nFactorial Incubator 2023 program. The goal of the game is to catch as many eggs as possible while avoiding bombs across three different levels of difficulty.

Developed in Python using the Pygame module, "Catch The Egg" features a user-friendly interface, responsive controls, and lively graphics and sound effects. I also integrated the Google Sheets API to store and rank player scores.

This project showcases my coding skills and creativity, representing a significant achievement as I completed my first-ever game in just three days. The code is structured using classes and functions for improved organization and readability, and detailed instructions on how to play the game can be found in the README file.

Feel free to download, modify, and share "Catch The Egg", following the terms and conditions of the MIT License. Enjoy the game!